# booking-service/src/main/resources/booking-openapi.yaml
openapi: 3.0.3
info:
  title: Room Booking API
  version: 1.0.0
  description: API для системы бронирования переговорных комнат
paths:
  #  Ниже добавлены 2 эндпоинта /api/v1/roomsdemo и /api/v1/bookingsdemo исключительно для проверки валидации по OpenApi мпецификации и проверки маскирования секретных полей
  /api/v1/roomsdemo:
    get:
      summary: Получить список всех переговорных
      parameters:
        - name: Authorization
          in: header
          required: true # Делаем заголовок обязательным
          description: "Bearer токен авторизации"
          schema:
            type: string
            # Задаем жесткое правило: токен должен быть строго от 50 до 100 символов
            minLength: 50
            maxLength: 100
      responses:
        '200':
          description: Успешный ответ со списком комнат
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/bookingsdemo:
    post:
      summary: Создать бронирование
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Бронирование успешно создано
        '409':
          description: Конфликт (комната уже занята на это время)
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '413':
          $ref: '#/components/responses/PayloadTooLargeError'
  /api/v1/rooms:
    get:
      summary: Получить список всех переговорных
      responses:
        '200':
          description: Успешный ответ со списком комнат
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/rooms/{roomId}/availability:
    get:
      summary: Получить расписание занятости комнаты на дату
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          description: Дата в формате YYYY-MM-DD
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/bookings:
    post:
      summary: Создать бронирование
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Бронирование успешно создано
        '409':
          description: Конфликт (комната уже занята на это время)
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '413':
          $ref: '#/components/responses/PayloadTooLargeError'

    get:
      summary: Получить список бронирований текущего пользователя
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/bookings/{bookingId}:
    delete:
      summary: Отменить бронирование
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Бронирование успешно отменено
        '404':
          description: Бронирование не найдено
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    Room:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Альфа"
        capacity:
          type: integer
          example: 10
        equipment:
          type: array
          items:
            type: string
          example: [ "Проектор", "Маркерная доска" ]

    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        isAvailable:
          type: boolean

    BookingRequest:
      type: object
      required:
        - roomId
        - userId
        - startTime
        - endTime
      properties:
        roomId:
          type: integer
          example: 1
        userId:
          type: integer
          example: 42
        startTime:
          type: string
          format: date-time
          example: "2026-04-12T10:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2026-04-12T11:00:00Z"

    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-04-12T10:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Schema request validation failed"
        details:
          type: array
          items:
            type: string
          example: [ "Field 'capacity' must be an integer" ]

  responses:
    BadRequestError:
      description: Ошибка валидации запроса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    InternalServerError:
      description: Нарушение контракта ответа бэкенда
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    PayloadTooLargeError:
      description: Размер тела запроса превышает допустимый лимит
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'